+++ src/jobs/xport.js
@@ -15,20 +15,20 @@ if (config.cache !== 'local') {
 const GeoXForm = require('geo-xform')
 const _ = require('highland')
 
-function xport (options, done) {
+function exportFile (options, done) {
   // TODO ya gotta fix this d00d
   if (!options.filePath) options.filePath = path.join('files', `/${options.id}_${options.layer || 0}`, options.key)
   const fileName = `${options.name}.geojson`
   options.geojsonPath = path.join(options.filePath, fileName)
-  if (options.format === 'geojson') return xformGeojson(options, done)
+  if (options.format === 'geojson') return exportGeojson(options, done)
   koop.files.exists(options.filePath, fileName, exists => {
     log.info('GeoJSON exists:', exists, options)
-    if (exists) return xformOnly(options, done)
-    xformAndSave(options, done)
+    if (exists) exportFromS3(options, done)
+    else exportFromCache(options, done)
   })
 }
 
-function xformOnly (options, done) {
+function exportFromS3 (options, done) {
   const fileName = options.geojsonPath.replace(/(?!\S+\.)geojson/, options.format)
   const transform = GeoXForm.createStream(options.format, options)
   const output = koop.files.createWriteStream(fileName)
@@ -45,21 +45,12 @@ function xformOnly (options, done) {
   .on('finish', () => done())
 }
 
-function xformAndSave (options, done) {
+function exportFromCache (options, done) {
   const cacheStream = createCacheStream(options)
-  const geojsonOut = koop.files.createWriteStream(options.geojsonPath)
   const fileName = `${options.name}.${options.format}`
   const output = koop.files.createWriteStream(path.join(options.filePath, fileName))
 
-  // we can pipe the geojson straight to S3 while transformations are in progress
-  // that way we can reuse the geojson rather than hitting the database again
   cacheStream
-  .fork()
-  .pipe(geojsonOut)
-  .on('error', e => log.error(e))
-
-  cacheStream
-	.fork()
   .pipe(GeoXForm.createStream(options.format, options))
   .on('error', e => done(e))
   .on('log', l => log[l.level](l.message))
@@ -70,7 +61,7 @@ function xformAndSave (options, done) {
   .on('finish', () => done())
 }
 
-function xformGeojson (options, done) {
+function exportGeojson (options, done) {
   createCacheStream(options)
   .on('log', l => log[l.level](l.message))
   .on('error', e => done(e))
@@ -92,4 +83,4 @@ function createCacheStream (options) {
   return output
 }
 
-module.exports = xport
+module.exports = exportFile