+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* `Rename xport to ExportFile`
+* Remove simeltaneous geojson + transform writing
+* Refactor exortFile internals
+
 ## [0.0.4] - 2016-01-26
 ### Changed
 * Use fork instead of observe to prevent memory issues