+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [0.1.0] - 2016-01-26
 ### Changed
 * `Rename xport to ExportFile`
 * Remove simeltaneous geojson + transform writing
@@ -24,6 +24,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * First release
 
+[0.1.0]: https://github.com/koopjs/koop-worker/compare/v0.0.4..v0.1.0
 [0.0.4]: https://github.com/koopjs/koop-worker/compare/v0.0.3..v0.0.4
 [0.0.3]: https://github.com/koopjs/koop-worker/compare/v0.0.2..v0.0.3
 [0.0.2]: https://github.com/koopjs/koop-worker/compare/v0.0.1..v0.0.2